extract time entry form into component
